src/requirements.txt
spotipy
python-dotenv
neo4j